refactor: 모임 신청 관련 Enum들을 Party 도메인 아래로 변경
