+++ src/main/java/umc/cockple/demo/domain/party/enums/RequestAction.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.party.enums;
 
 public enum RequestAction {
     //승인, 거절 요청
