+++ src/main/java/umc/cockple/demo/domain/party/repository/PartyJoinRequestRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.party.domain.Party;
 import umc.cockple.demo.domain.party.domain.PartyJoinRequest;
-import umc.cockple.demo.global.enums.RequestStatus;
+import umc.cockple.demo.domain.party.enums.RequestStatus;
 
 public interface PartyJoinRequestRepository extends JpaRepository<PartyJoinRequest, Long> {
     boolean existsByPartyAndMemberAndStatus(Party party, Member member, RequestStatus requestStatus);
